+++ RELEASE-NOTES.md
@@ -1,5 +1,12 @@
 # Release Notes
 
+## 1.0.0-RC01
+
+- Major project refactoring and cleanup
+- The core utilities that have no external dependencies are moved to the `core` module
+- The config utilities framework is moved to the `config-z` module
+- Added more utilities, for `Future`, `Either` and guarding code through `Bracket`s
+
 
 ## 0.2.0
   - `Configurator`s can also be used as `Extractor`s
@@ -8,7 +15,7 @@
   - Added extractors for time related properties: `Duration`, `Timestamp`, `Date`, `LocalDateTime` and `LocalDate`
 
 ## 0.1.0
-  - [`Configurator`](docs/configuration-framework.md)s framework
-  - [`byteable`](src/main/scala/org/tupol/utils/byteable.scala) to convert various data types into array of bytes
-  - [`utils`](src/main/scala/org/tupol/utils/utils.scala) with various utilities for `Map`s, `Product`s and `Try`s
+  - `Configurator`s framework
+  - `byteable` to convert various data types into array of bytes
+  - `utils` with various utilities for `Map`s, `Product`s and `Try`s